+++ feature/src/main/java/com/zcard/feature/cardeditor/CardEditorFragment.kt
@@ -33,6 +33,8 @@ import com.zcard.card.databinding.FragmentCardEditorBinding
 import com.zcard.designsystem.theme.ZCardTheme
 import com.zcard.domain.model.UnityMessage
 import com.zcard.domain.model.UnityEventType
+import com.zcard.feature.MainSideEffect
+import com.zcard.feature.MainViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import kotlin.getValue
