+++ feature/src/main/res/layout/activity_card_editor.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zcard.feature.cardeditor.MainActivity">
+    tools:context="com.zcard.feature.MainActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/unity_container"
